requirements.txt
opencv-contrib-python==4.8.1.78
pillow==10.0.1
numpy==2.0.2